Simplify/improve sink* and source* files
I still need to make it so the *raise files won't go over 100%; this is
such a huge oversight for the volume setting ability of pactl(1), and
there's no decent way to get the volume of the default sink/source.

The code used to work, but I broke it at some point, so it hasn't been
working for quite some time.