extended globbing version
while splitting font size and value, instead of splitting by spaces, try to target numbers with extended globbing. that way you'll preserve the number of spaces and only edit the actual values. i am not exactly sure this is what you wanted, i've seen the video and understood what you were trying to do, and converted all regex stuff into extended globbing and when i saw your benchmark test changes, it seemed you were trying to do the same thing as the changes i made, but again, not sure and my mind is asleep. if this wasn't the point, let me know and ill work on it with you.