Run 'autorun' if found and executable
This script is a loop script for notifying without the messy nonsense
that can be Cron with notifications. I've tried setting `DISPLAY` and
`XDG_RUNTIME_DIR` variables to the correct value from both the user's
crontab and the scripts it runs, but it sometimes works and sometimes
doesn't, so, I give up -- just going to do it the tried and true way!
For simple terminal-only stuff, Cron is great, but for things like
notifications? It's horrendous.