Add script to toggle middle mouse button
I don't personally need this, but it might wind up being useful for
some reason. Mainly, it's just something I see a lot of people wanting
to do in Linux, so at least I can point them to this script.

It does first require the user enter `xinput` to themselves discern the
correct device ID for their mouse, but it's fairly straightforward.