Optimize (no REGEX) and correct (globs alert)
Quote-protected variables on right side of string equality check,
in-case for some reason the variable expanded to something with glob
pattern matching syntax the shell would interpret.

Removed REGEX and used extended globbing, since this is by default
enabled with the `[[` shell keyword, thankfully. This is a slight
optimization which should make a bit of difference, given the rate at
which these scripts can be called.